# Parallelise playwright tests in axe-playwright action