--- a/packages/design-tokens/src/tokens/functional/colors/global.json
+++ b/packages/design-tokens/src/tokens/functional/colors/global.json
@@ -2,121 +2,121 @@
   "color": {
     "success": {
       "fg": {
-        "value": "var(--base-color-scale-green-5)",
-        "dark": "var(--base-color-scale-green-3)"
+        "value": "hotpink",
+        "dark": "hotpink"
       },
       "emphasis": {
-        "value": "var(--base-color-scale-green-4)",
-        "dark": "var(--base-color-scale-green-5)"
+        "value": "hotpink",
+        "dark": "hotpink"
       },
       "muted": {
-        "value": "#4bc36b99",
-        "dark": "#2e9e4299"
+        "value": "hotpink",
+        "dark": "hotpink"
       },
       "subtle": {
-        "value": "var(--base-color-scale-green-0)",
-        "dark": "#2e9e42d9"
+        "value": "hotpink",
+        "dark": "hotpink"
       }
     },
     "error": {
       "fg": {
-        "value": "var(--base-color-scale-red-5)",
-        "dark": "var(--base-color-scale-red-4)"
+        "value": "hotpink",
+        "dark": "hotpink"
       },
       "emphasis": {
-        "value": "var(--base-color-scale-red-5)",
-        "dark": "var(--base-color-scale-red-5)"
+        "value": "hotpink",
+        "dark": "hotpink"
       },
       "muted": {
-        "value": "#ff878599",
-        "dark": "#f8524999"
+        "value": "hotpink",
+        "dark": "hotpink"
       },
       "subtle": {
-        "value": "var(--base-color-scale-red-0)",
-        "dark": "#f85249d9"
+        "value": "hotpink",
+        "dark": "hotpink"
       }
     },
     "accent": {
       "primary": {
-        "value": "var(--base-color-scale-green-5)",
-        "dark": "var(--base-color-scale-green-3)"
+        "value": "hotpink",
+        "dark": "hotpink"
       },
       "secondary": {
-        "value": "var(--base-color-scale-yellow-5)",
-        "dark": "var(--base-color-scale-yellow-3)"
+        "value": "hotpink",
+        "dark": "hotpink"
       }
     },
     "text": {
       "default": {
-        "value": "var(--base-color-scale-gray-9)",
-        "dark": "var(--base-color-scale-white-0)"
+        "value": "hotpink",
+        "dark": "hotpink"
       },
       "muted": {
-        "value": "var(--base-color-scale-gray-6)",
-        "dark": "var(--base-color-scale-gray-3)"
+        "value": "hotpink",
+        "dark": "hotpink"
       },
       "subtle": {
-        "value": "var(--base-color-scale-gray-5)",
-        "dark": "var(--base-color-scale-gray-4)"
+        "value": "hotpink",
+        "dark": "hotpink"
       },
       "onEmphasis": {
-        "value": "var(--base-color-scale-white-0)",
-        "dark": "var(--base-color-scale-black-0)"
+        "value": "hotpink",
+        "dark": "hotpink"
       }
     },
     "neutral": {
       "emphasisPlus": {
-        "value": "var(--base-color-scale-gray-9)",
-        "dark": "var(--base-color-scale-gray-4)"
+        "value": "hotpink",
+        "dark": "hotpink"
       },
       "emphasis": {
-        "value": "var(--base-color-scale-gray-5)",
-        "dark": "var(--base-color-scale-gray-4)"
+        "value": "hotpink",
+        "dark": "hotpink"
       },
       "muted": {
-        "value": "#b7bfc8fe",
-        "dark": "#6e7681fe"
+        "value": "hotpink",
+        "dark": "hotpink"
       },
       "subtle": {
-        "value": "#e9edf1fe",
-        "dark": "#6e7681ff"
+        "value": "hotpink",
+        "dark": "hotpink"
       }
     },
     "canvas": {
       "default": {
-        "value": "var(--base-color-scale-white-0)",
-        "dark": "var(--base-color-scale-black-0)"
+        "value": "hotpink",
+        "dark": "hotpink"
       },
       "overlay": {
-        "value": "var(--base-color-scale-white-0)",
-        "dark": "var(--base-color-scale-gray-8)"
+        "value": "hotpink",
+        "dark": "hotpink"
       },
       "inset": {
-        "value": "var(--base-color-scale-gray-0)",
-        "dark": "var(--base-color-scale-black-0)"
+        "value": "hotpink",
+        "dark": "hotpink"
       },
       "subtle": {
-        "value": "var(--base-color-scale-gray-0)",
-        "dark": "var(--base-color-scale-gray-8)"
+        "value": "hotpink",
+        "dark": "hotpink"
       }
     },
     "border": {
       "default": {
-        "value": "var(--base-color-scale-gray-3)",
-        "dark": "var(--base-color-scale-gray-6)"
+        "value": "hotpink",
+        "dark": "hotpink"
       },
       "muted": {
-        "value": "#d3d9dffe",
-        "dark": "var(--base-color-scale-gray-7)"
+        "value": "hotpink",
+        "dark": "hotpink"
       },
       "subtle": {
-        "value": "#e9edf1fe",
-        "dark": "var(--base-color-scale-gray-8)"
+        "value": "hotpink",
+        "dark": "hotpink"
       }
     },
     "focus": {
-      "value": "var(--base-color-scale-blue-5)",
-      "dark": "var(--base-color-scale-blue-3)"
+      "value": "hotpink",
+      "dark": "hotpink"
     }
   }
 }