--- a/packages/e2e/scripts/playwright/axe-clean.spec.ts
+++ b/packages/e2e/scripts/playwright/axe-clean.spec.ts
@@ -115,6 +115,8 @@ const storybookRoutes = Object.values((IndexData as StoryIndex).entries)
     return !testsToSkip.includes(id)
   })
 
+test.describe.configure({mode: 'parallel'})
+
 for (const story of storybookRoutes) {
   // eslint-disable-next-line i18n-text/no-en
   describe(`Web page accessibility test for ${story.name} - ${story.story}`, () => {