## Summary

Inspired by the recent improvements to our VRT tests, this PR parallelises our axe-playwright tests, used in our UI test / Accessibility action.

| Previous time taken | Time taken now | Improvement |
| --- | --- | --- | 
| 22 mins | 7 mins | **15 mins faster** 🎉 ⚡ |

See the Playwright docs on the change here: https://playwright.dev/docs/api/class-test#test-describe-configure

## Testing

You can see from the checks associated with this PR that the action is completing successfully.

In an earlier commit (f80de414a31178cc0fcb6ded7fff9ef6f320e475) I intentionally triggered axe violations to test that they would still be reported. As you can see from [this run](https://github.com/primer/brand/actions/runs/15461299873/job/43523045254?pr=1040) they were reported successfully.

All of the axe tests run independently in our test suite with no shared state, so it seems reasonable that we should be able to parallelise them without any issues.

## What should reviewers focus on?

- If/when this PR is merged, keep an eye on the UI test / Accessibility action for your next few PRs to make sure it continues to work correctly.

## Steps to test:

1. Run the UI test / Accessibility action



## Reviewer checklist:

- [ ] Check that pull request and proposed changes adhere to our [contribution guidelines](../../CONTRIBUTING.md) and [code of conduct](../../CODE_OF_CONDUCT.md)
- [ ] Check that tests prove the feature works and covers both happy and unhappy paths
- [ ] Check that there aren't other open Pull Requests for the same update/change

## Screenshots:

> Please try to provide before and after screenshots or videos

<table>
<tr>
<th> Before</th> <th> After </th>
</tr>
<tr>
<td valign="top">

![image](https://github.com/user-attachments/assets/43a44d80-f9c8-486d-9acc-b935d33a2dce)


 </td>
<td valign="top">

![image](https://github.com/user-attachments/assets/cc9b6f6f-ec02-4a30-9fc6-99147abbf244)

</td>
</tr>
</table>